requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
requests==2.32.4
pypdf==6.0.0
python-docx==1.2.0
